Refactors snapshot implementation
