+++ src/Management/Mapper/EntrySnapshot.php
@@ -31,7 +31,7 @@ class EntrySnapshot extends BaseMapper
 
         return $this->hydrate(ResourceClass::class, [
             'sys' => new SystemProperties($data['sys']),
-            'entry' => $this->builder->build($data['snapshot']),
+            'snapshot' => $this->builder->build($data['snapshot']),
         ]);
     }
 }
